+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- feature: Add `canRun` boolean to onChange menuItems object.
 
 ## [5.1.2] - 2018-11-12
 - update: headerIds more cleanly handles additional whitespace.